First round of tweaks
* Adds targeting data component.
* Remove the entity pool for zombies.
* Remove the dependency on array indices.
* Replace activity checks with selection tag.
(aka the empty components 'Active' and 'Infected')
* Fix and Optimized the entity navigation movement systems.

Note: Job Execution is not optimal atm, with large sync
points and idle threads there is no need to make it optimal.
Best to wait till more jobs are implemented to fill the space.

Note: Swapping out MeshInstanceRenderer is a bit of an issue,
but the biggest slow down atm is the command buffer playback.
The amount of human infections and disabled zombies is to high
and all is being pushed on to the main thread to sync per frame.