Remove command buffer requirement
* Turn HumanInfectionSystem in to JobComponentSystem
* Logic (Human / Zombie) components are combined using actor
* Adjust other systems to only run on actor they where designed for.

Note: This breaks some of ECS principles but it seems the fastest way to update a large amount of entities without slowing down the main thread. At this point might as well combine some of the systems to reduce the amount of times the world needs to loop over all actor entities.